Finish

+++ app/src/main/java/com/mapl/tinkoff/CardInfo.java
@@ -1,11 +1,11 @@
 package com.mapl.tinkoff;
 
 public class CardInfo {
-    long id;
+    Integer id;
     long milliseconds;
     String text;
 
-    public CardInfo(long id, long milliseconds, String text) {
+    public CardInfo(Integer id, long milliseconds, String text) {
         this.id = id;
         this.milliseconds = milliseconds;
         this.text = text;
